HeapSort
Heap sort is a comparison-based sorting method based on the Binary Heap Data Structure. It may be thought of as an optimization over selection sort, in which we first discover the maximum (or minimum) element and exchange it with the last (or first). We repeat the process with the remaining components. In Heap Sort, we utilize Binary Heap to swiftly discover and shift the maximum element in O(Log n) rather than O(n), resulting in an O(n Log n) time complexity.